Show the playlist in the web UI
Recently I've been playing music and Oiva wants to jump to a
specific song in the playlist - I know it is there, but clicking
"Next", "Next", "Next" until I hit it is frustrating.

This commit adds the current contents of the playlist to the foot
of the display, and clicking a specific entry will start playing
it.